refactor: userId를 Int로 수정

+++ android/shared/src/commonMain/kotlin/com/yeobaek/core/analytics/AnalyticsTracker.kt
@@ -14,7 +14,7 @@ class AnalyticsTracker(
         client.screen(screenName)
     }
 
-    fun identify(userId: Long) {
+    fun identify(userId: Int) {
         client.identify(userId.toString())
     }
 }
